🚀 macOS Port Progress: Major Socket API Enhancement & Error Reduction (74→36)
✅ Socket API Refinement:
- Enhanced Windows Socket API compatibility with closesocket guards
- Complete WSAE error constants (WSAEWOULDBLOCK, WSAEINVAL, etc.)
- Fixed socket function redefinition conflicts in WWDownload/winsock.h

✅ Windows API Enhancement:
- Added IN/OUT SAL annotations for SNMP and Windows functions
- Resolved SNMP type conflicts (AsnObjectIdentifier, RFC1157VarBindList)
- Improved header coordination between win32_compat.h and existing headers

✅ 64-bit Compatibility:
- Fixed pointer-to-integer casting using intptr_t for GameSpy callbacks
- Applied ICMP structure protection with cross-platform guards
- Enhanced calling convention macros (WINAPI, __stdcall, FAR)

✅ Core Libraries Success:
- Multiple WW libraries now compile without errors
- Demonstrated successful build of core_wwlib, core_wwmath, core_wwdebug
- Systematic windows.h replacement with conditional win32_compat.h inclusion

🎯 Progress: Reduced compilation errors from 74 to 36 in generalszh target
📝 Documentation: Updated MACOS_PORT.md and README.md with current session progress

+++ GeneralsMD/Code/GameEngine/Source/GameNetwork/GameSpy/MainMenuUtils.cpp
@@ -313,7 +313,7 @@ static void queuePatch(Bool mandatory, AsciiString downloadURL)
 static GHTTPBool motdCallback( GHTTPRequest request, GHTTPResult result,
 															char * buffer, GHTTPByteCount bufferLen, void * param )
 {
-	Int run = (Int)param;
+	Int run = (intptr_t)param;
 	if (run != timeThroughOnline)
 	{
 		DEBUG_CRASH(("Old callback being called!"));
@@ -353,7 +353,7 @@ static GHTTPBool motdCallback( GHTTPRequest request, GHTTPResult result,
 static GHTTPBool configCallback( GHTTPRequest request, GHTTPResult result,
 																char * buffer, GHTTPByteCount bufferLen, void * param )
 {
-	Int run = (Int)param;
+	Int run = (intptr_t)param;
 	if (run != timeThroughOnline)
 	{
 		DEBUG_CRASH(("Old callback being called!"));
@@ -418,7 +418,7 @@ static GHTTPBool configCallback( GHTTPRequest request, GHTTPResult result,
 static GHTTPBool configHeadCallback( GHTTPRequest request, GHTTPResult result,
 																		char * buffer, GHTTPByteCount bufferLen, void * param )
 {
-	Int run = (Int)param;
+	Int run = (intptr_t)param;
 	if (run != timeThroughOnline)
 	{
 		DEBUG_CRASH(("Old callback being called!"));
